feat: upgrade Next.js to 16.2.4, integrate Storybook 8.6, and refactor ProsperLayout for better modularity
