+++ apps/storybook/.storybook/main.ts
@@ -3,16 +3,14 @@ import path from "path";
 
 const config: StorybookConfig = {
   stories: [
-    "../src/**/*.stories.@(js|jsx|ts|tsx)",
     "../../../packages/ui/src/**/*.stories.@(js|jsx|ts|tsx)",
-    "../../../apps/web-main/app/**/*.stories.@(js|jsx|ts|tsx)",
-    "../../../apps/web-prosper/app/**/*.stories.@(js|jsx|ts|tsx)",
-    "../../../apps/web-prosper/components/**/*.stories.@(js|jsx|ts|tsx)",
+    "../../../apps/web-prosper/{app,components,layout,lib,hooks}/**/*.stories.@(js|jsx|ts|tsx)",
   ],
   addons: [
     "@storybook/addon-links",
     "@storybook/addon-essentials",
     "@storybook/addon-interactions",
+    "@storybook/addon-viewport",
   ],
   framework: {
     name: "@storybook/experimental-nextjs-vite",
@@ -31,6 +29,8 @@ const config: StorybookConfig = {
       config.resolve.alias = {
         ...config.resolve.alias,
         "@": path.resolve(__dirname, "../src"),
+        "next/config": path.resolve(__dirname, "./next-config-mock.js"),
+        "storybook/test": "@storybook/test",
       };
     }
     return config;
