+++ apps/storybook/.storybook/next-config-mock.js
@@ -1,0 +1,9 @@
+export default () => ({
+  publicRuntimeConfig: {},
+  serverRuntimeConfig: {},
+});
+export const getConfig = () => ({
+  publicRuntimeConfig: {},
+  serverRuntimeConfig: {},
+});
+export const setConfig = () => {};
