+++ apps/storybook/.storybook/preview.tsx
@@ -1,5 +1,6 @@
 import React, { useEffect } from "react";
 import type { Preview } from "@storybook/react";
+import { INITIAL_VIEWPORTS } from "@storybook/addon-viewport";
 import "./globals.css";
 
 /**
@@ -64,6 +65,9 @@ function ThemeDecorator({
 
 const preview: Preview = {
   parameters: {
+    viewport: {
+      viewports: INITIAL_VIEWPORTS,
+    },
     controls: {
       matchers: {
         color: /(background|color)$/i,
