+++ apps/storybook/package.json
@@ -17,19 +17,21 @@
   },
   "devDependencies": {
     "@ously/tsconfig": "workspace:*",
-    "@storybook/addon-essentials": "8.6.14",
-    "@storybook/addon-interactions": "8.6.14",
-    "@storybook/addon-links": "8.6.14",
-    "@storybook/blocks": "8.6.14",
-    "@storybook/experimental-nextjs-vite": "8.6.14",
-    "@storybook/react": "8.6.14",
-    "@storybook/test": "8.6.14",
+    "@storybook/addon-essentials": "catalog:",
+    "@storybook/addon-interactions": "catalog:",
+    "@storybook/addon-links": "catalog:",
+    "@storybook/addon-viewport": "catalog:",
+    "@storybook/blocks": "catalog:",
+    "@storybook/experimental-nextjs-vite": "catalog:",
+    "@storybook/react": "catalog:",
+    "@storybook/test": "catalog:",
     "@types/react": "catalog:",
     "@types/react-dom": "catalog:",
     "autoprefixer": "catalog:",
-    "next": "^15.1.0",
+    "next": "^16.2.4",
     "postcss": "catalog:",
-    "storybook": "8.6.14",
+    "prop-types": "^15.8.1",
+    "storybook": "catalog:",
     "tailwindcss": "catalog:",
     "typescript": "catalog:",
     "vite": "^5.2.11"
