+++ apps/web-prosper/app/globals.css
@@ -1,5 +1,5 @@
-@import "tw-animate-css";
-@import "shadcn/tailwind.css";
+@import "../../../node_modules/tw-animate-css/dist/tw-animate.css";
+@import "../../../node_modules/shadcn/dist/tailwind.css";
 @import "../../../packages/ui/src/themes.css";
 @tailwind base;
 @tailwind components;
