+++ apps/web-prosper/components/__stories__/MockCard.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import { MockCard } from "./MockCard";
+import { MockCard } from "../MockCard";
 
 const meta = {
   title: "Prosper/MockCard",
@@ -7,7 +7,8 @@ const meta = {
   tags: ["autodocs"],
   args: {
     title: "Prosper Feature",
-    description: "This is a mock component specific to the Prosper application.",
+    description:
+      "This is a mock component specific to the Prosper application.",
   },
 } satisfies Meta<typeof MockCard>;
