+++ apps/web-prosper/package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "dev": "next dev",
     "build": "npx @cloudflare/next-on-pages",
-    "build:next": "next build",
+    "build:next": "next build --webpack",
     "start": "next start",
     "lint": "next lint",
     "type-check": "tsc --noEmit"
@@ -17,7 +17,8 @@
     "@phosphor-icons/react": "^2.1.10",
     "class-variance-authority": "catalog:",
     "clsx": "catalog:",
-    "next": "14.2.3",
+    "lucide-react": "catalog:",
+    "next": "16.2.4",
     "radix-ui": "^1.4.3",
     "react": "catalog:",
     "react-dom": "catalog:",
@@ -28,6 +29,7 @@
   "devDependencies": {
     "@cloudflare/next-on-pages": "^1.13.16",
     "@ously/tsconfig": "workspace:*",
+    "@storybook/react": "catalog:",
     "@types/node": "^20",
     "@types/react": "catalog:",
     "@types/react-dom": "catalog:",
