+++ package.json
@@ -10,14 +10,18 @@
     "storybook": "turbo dev --filter=storybook"
   },
   "devDependencies": {
-    "turbo": "^2.7.2",
+    "next": "16.2.4",
     "prettier": "^3.0.0",
+    "shadcn": "^4.6.0",
+    "turbo": "^2.7.2",
+    "tw-animate-css": "^1.4.0",
     "typescript": "^5.4.5"
   },
   "packageManager": "pnpm@10.30.3",
   "pnpm": {
     "overrides": {
-      "webpack": "^5.91.0"
+      "webpack": "^5.91.0",
+      "vite-plugin-storybook-nextjs": "^3.2.4"
     }
   }
 }
