+++ packages/ui/package.json
@@ -2,6 +2,7 @@
   "name": "@ously/ui",
   "version": "0.0.0",
   "private": true,
+  "type": "module",
   "main": "./src/index.ts",
   "types": "./src/index.ts",
   "exports": {
@@ -23,7 +24,7 @@
     "test:watch": "vitest"
   },
   "peerDependencies": {
-    "next": "^14.0.0 || ^15.0.0",
+    "next": "^16.0.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1"
   },
